[ai] Make ExamplesTest run without runtime reflection
Prompt:

Make class Lox resettable - so that the error state and the
interpreter can be reset to a fresh state by call to a public method
reset(). Call this method before every runFile call in
ExamplesTest. No reflection should be used to achieve this, but
instead you can change finality and/or visibility of fields and
methods of Lox class.